Move the code of `ast_step0` to `parse`
