+++ src/lib.rs
@@ -1,4 +1,3 @@
-mod ast_step0;
 mod ast_step1;
 mod ast_step2;
 mod ast_step3;
